refactor: validateJson
