--- src/pages/tools/json/validateJson/servcie.ts
@@ -1,14 +1,0 @@
-export const validateJson = (
-  input: string
-): { valid: boolean; error?: string } => {
-  try {
-    JSON.parse(input);
-    return { valid: true };
-  } catch (error) {
-    if (error instanceof SyntaxError) {
-      const message = error.message;
-      return { valid: false, error: error.message };
-    }
-    return { valid: false, error: 'Unknown error occurred' };
-  }
-};
