+++ src/pages/tools/json/validateJson/service.ts
@@ -1,0 +1,13 @@
+export const validateJson = (
+  input: string
+): { valid: boolean; error?: string } => {
+  try {
+    JSON.parse(input);
+    return { valid: true };
+  } catch (error) {
+    if (error instanceof SyntaxError) {
+      return { valid: false, error: error.message };
+    }
+    return { valid: false, error: 'Unknown error occurred' };
+  }
+};
